refactor: Move simplifyTransaction utility to services and update imports in related components
